CUDA compiles with many warnings

+++ src/chimaera_runtime.cc
@@ -167,7 +167,7 @@ void Runtime::InitSharedMemoryGpu() {
     ChiShm *header = main_alloc_->GetCustomHeader<ChiShm>();
     header->node_id_ = CHI_RPC->node_id_;
     header->unique_ =
-        (1 << 32); // TODO(llogan): Make a separate unique for gpus
+        (((u64)1) << 32);  // TODO(llogan): Make a separate unique for gpus
     header->num_nodes_ = server_config_.rpc_.host_names_.size();
   }
 #endif
@@ -185,7 +185,7 @@ void Runtime::InitSharedMemoryGpu() {
     ChiShm *header = main_alloc_->GetCustomHeader<ChiShm>();
     header->node_id_ = CHI_RPC->node_id_;
     header->unique_ =
-        (((u64)1) << 32); // TODO(llogan): Make a separate unique for gpus
+        (((u64)1) << 32);  // TODO(llogan): Make a separate unique for gpus
     header->num_nodes_ = server_config_.rpc_.host_names_.size();
   }
 #endif
@@ -205,4 +205,4 @@ void Runtime::RunDaemon() {
 /** Stop the Hermes core Daemon */
 void Runtime::StopDaemon() { CHI_WORK_ORCHESTRATOR->FinalizeRuntime(); }
 
-} // namespace chi
+}  // namespace chi
